Add retry to juju status.
