+++ cmd/juju/status/export_test.go
@@ -3,8 +3,17 @@
 
 package status
 
-import "github.com/juju/cmd"
+import (
+	"github.com/juju/cmd"
+
+	"github.com/juju/juju/cmd/modelcmd"
+)
 
 func NewTestStatusHistoryCommand(api HistoryAPI) cmd.Command {
 	return &statusHistoryCommand{api: api}
+}
+
+func NewTestStatusCommand(api statusAPI, clock Clock) cmd.Command {
+	return modelcmd.Wrap(
+		&statusCommand{api: api, clock: clock})
 }
